Moving Column class into dedicated source file
